refactor(project): package with initial A - Refactor package.json for package with initial A and its dependencies
