+++ packages/affix/package.json
@@ -1,12 +1,16 @@
 {
   "name": "@element-plus/affix",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "main": "dist/index.js",
   "license": "MIT",
   "peerDependencies": {
     "vue": "3.1.1"
   },
+  "dependencies": {
+    "@element-plus/utils": "^0.0.1"
+  },
   "devDependencies": {
-    "@vue/test-utils": "^2.0.0-beta.3"
+    "@vue/test-utils": "^2.0.0-beta.3",
+    "@element-plus/test-utils": "^0.0.1"
   }
 }
