+++ packages/alert/package.json
@@ -1,10 +1,13 @@
 {
   "name": "@element-plus/alert",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "main": "dist/index.js",
   "license": "MIT",
   "peerDependencies": {
     "vue": "3.1.1"
+  },
+  "dependencies": {
+    "@element-plus/utils": "^0.0.1"
   },
   "devDependencies": {
     "@vue/test-utils": "^2.0.0-beta.3"
