+++ packages/aside/package.json
@@ -6,6 +6,9 @@
   "peerDependencies": {
     "vue": "3.1.1"
   },
+  "dependencies": {
+    "@element-plus/utils": "^0.0.1"
+  },
   "devDependencies": {
     "@vue/test-utils": "^2.0.0-beta.3"
   }
