+++ packages/avatar/package.json
@@ -10,6 +10,6 @@
     "@vue/test-utils": "^2.0.0-beta.3"
   },
   "dependencies": {
-    "@element-plus/utils": "^0.0.0"
+    "@element-plus/utils": "^0.0.1"
   }
 }
