+++ packages/container/__tests__/container.spec.ts
@@ -2,7 +2,7 @@ import { mount } from '@vue/test-utils'
 import Container from '../src/container.vue'
 import Header from '../src/header.vue'
 import Main from '../src/main.vue'
-import Aside from '../src/aside.vue'
+import Aside from '../../aside/src/aside.vue'
 import Footer from '../src/footer.vue'
 import { getCssVariable } from '@element-plus/test-utils'
