+++ packages/scrollbar/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@element-plus/scrollbar",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "main": "dist/index.js",
   "license": "MIT",
   "peerDependencies": {
