+++ packages/utils/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@element-plus/utils",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "license": "MIT",
   "dependencies": {
     "resize-observer-polyfill": "^1.5.1"
